# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Suivi Académique",
        "description": "Documentation de l'API pour le suivi académique",
        "contact": {
            "email": "contact@suivi-academique.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur API"
        }
    ],
    "paths": {
        "/api/ec": {
            "get": {
                "tags": [
                    "ECs"
                ],
                "summary": "Obtenir la liste des EC",
                "description": "Retourne la liste de toutes les EC avec leur UE associée",
                "operationId": "getECList",
                "responses": {
                    "200": {
                        "description": "Opération réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "code_ec": {
                                                "type": "integer"
                                            },
                                            "label_ec": {
                                                "type": "string"
                                            },
                                            "description_ec": {
                                                "type": "string"
                                            },
                                            "nb_heures_ec": {
                                                "type": "integer"
                                            },
                                            "nb_credits_ec": {
                                                "type": "integer"
                                            },
                                            "code_ue": {
                                                "type": "string"
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "ECs"
                ],
                "summary": "Créer un nouvel EC",
                "description": "Crée un EC et retourne un message de confirmation",
                "operationId": "storeEC",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "label_ec",
                                    "description_ec",
                                    "nb_heures_ec",
                                    "nb_credits_ec",
                                    "code_ue"
                                ],
                                "properties": {
                                    "label_ec": {
                                        "type": "string",
                                        "minLength": 3
                                    },
                                    "description_ec": {
                                        "type": "string"
                                    },
                                    "nb_heures_ec": {
                                        "type": "integer"
                                    },
                                    "nb_credits_ec": {
                                        "type": "integer"
                                    },
                                    "code_ue": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "EC créé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/ec/{codeEC}": {
            "get": {
                "tags": [
                    "ECs"
                ],
                "summary": "Obtenir un EC",
                "description": "Retourne les détails d'un EC",
                "operationId": "showEC",
                "parameters": [
                    {
                        "name": "codeEC",
                        "in": "path",
                        "description": "Code de l'EC",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Opération réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "EC non trouvé"
                    }
                }
            },
            "put": {
                "tags": [
                    "ECs"
                ],
                "summary": "Modifier un EC",
                "description": "Modifie un EC existant",
                "operationId": "updateEC",
                "parameters": [
                    {
                        "name": "codeEC",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "label_ec": {
                                        "type": "string"
                                    },
                                    "description_ec": {
                                        "type": "string"
                                    },
                                    "nb_heures_ec": {
                                        "type": "integer"
                                    },
                                    "nb_credits_ec": {
                                        "type": "integer"
                                    },
                                    "code_ue": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "EC mis à jour"
                    },
                    "404": {
                        "description": "EC non trouvé"
                    }
                }
            },
            "delete": {
                "tags": [
                    "ECs"
                ],
                "summary": "Supprimer un EC",
                "description": "Supprime un EC existant",
                "operationId": "deleteEC",
                "parameters": [
                    {
                        "name": "codeEC",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "EC supprimé"
                    },
                    "404": {
                        "description": "EC non trouvé"
                    }
                }
            }
        },
...
    },
...
}